+++ packages/apollo-angular-components/projects/angular/components/ui-grid/src/body/ui-grid-loading.directive.ts
@@ -5,11 +5,11 @@ import {
 } from '@angular/core';
 
 @Directive({
-  selector: '[uiGridLoading], ui-grid-loading',
+    selector: '[uiGridLoading], ui-grid-loading',
 })
 export class UiGridLoadingDirective {
-  @ContentChild(TemplateRef, {
-    static: true,
-  })
-  public html?: TemplateRef<any>;
+    @ContentChild(TemplateRef, {
+        static: true,
+    })
+    html?: TemplateRef<any>;
 }