+++ packages/apollo-angular-components/projects/angular/components/ui-grid/src/managers/resize/strategies/immediate-neighbour-halt-resizer.ts
@@ -24,7 +24,7 @@ export class ImmediateNeighbourHaltResizer<T extends IGridDataEntry> extends Res
         }
 
         return true;
-    }
+    };
 
     protected _resizeLeftFilter = (state: IResizeEvent<T>) => {
         if (state.current.direction !== ResizeDirection.Left) { return true; }
@@ -36,7 +36,7 @@ export class ImmediateNeighbourHaltResizer<T extends IGridDataEntry> extends Res
         }
 
         return true;
-    }
+    };
 
     protected _resizeRightFilter = (state: IResizeEvent<T>) => {
         if (state.current.direction !== ResizeDirection.Right) { return true; }
@@ -50,7 +50,7 @@ export class ImmediateNeighbourHaltResizer<T extends IGridDataEntry> extends Res
         }
 
         return true;
-    }
+    };
 
     protected _stateUpdate = (state: IResizeEvent<T>) => {
         if (state.current.direction === ResizeDirection.Left) {
@@ -70,5 +70,5 @@ export class ImmediateNeighbourHaltResizer<T extends IGridDataEntry> extends Res
             // add the stolen width to the actively resized column
             this._applyOffsetFor(state.current.resized, offset);
         }
-    }
+    };
 }