+++ packages/apollo-angular-components/projects/angular/components/ui-grid/src/managers/selection-manager.ts
@@ -21,26 +21,27 @@ import {
  * @ignore
  */
 export class SelectionManager<T extends IGridDataEntry> {
-    public get selected(): T[] {
+    get selected(): T[] {
         return Array.from(this._selection.values());
     }
 
-    public get selectionSnapshot(): T[] {
+    get selectionSnapshot(): T[] {
         return Array.from(this._selectionSnapshot.values());
     }
 
-    public get difference(): ISelectionDiff<T> {
+    get difference(): ISelectionDiff<T> {
         return {
             add: differenceBy(this.selected, this.selectionSnapshot, 'id'),
             remove: differenceBy(this.selectionSnapshot, this.selected, 'id'),
         };
     }
 
-    public changed$ = new Subject<SelectionChange<T>>();
+    changed$ = new Subject<SelectionChange<T>>();
 
     private _hasValue$ = new BehaviorSubject(false);
 
-    public hasValue$ = this._hasValue$.pipe(distinctUntilChanged());
+    // eslint-disable-next-line @typescript-eslint/member-ordering
+    hasValue$ = this._hasValue$.pipe(distinctUntilChanged());
 
     private _selection = new Map<number | string, T>();
 
@@ -55,45 +56,46 @@ export class SelectionManager<T extends IGridDataEntry> {
         private _emitChanges = true,
     ) {
 
-        if (initiallySelectedValues && initiallySelectedValues.length) {
+        if (initiallySelectedValues?.length) {
             initiallySelectedValues.forEach(value => this._markSelected(value));
             this._selectedToEmit.length = 0;
         }
     }
 
-    public select = (...values: T[]): void =>
-        this._updateState(this._markSelected, values)
+    select = (...values: T[]): void =>
+        this._updateState(this._markSelected, values);
 
-    public deselect = (...values: T[]): void =>
-        this._updateState(this._unmarkSelected, values)
+    deselect = (...values: T[]): void =>
+        this._updateState(this._unmarkSelected, values);
 
-    public toggle(value: T): void {
+    toggle(value: T): void {
+        // eslint-disable-next-line @typescript-eslint/no-unused-expressions
         this.isSelected(value) ? this.deselect(value) : this.select(value);
     }
 
-    public clear(): void {
+    clear(): void {
         this._selection.forEach(v => this._unmarkSelected(v));
         this._selection.clear();
         this._emitChangeEvent();
     }
 
-    public isSelected(value: T): boolean {
+    isSelected(value: T): boolean {
         return this._selection.has(value.id);
     }
 
-    public isEmpty(): boolean {
+    isEmpty(): boolean {
         return this._selection.size === 0;
     }
 
-    public hasValue(): boolean {
+    hasValue(): boolean {
         return !this.isEmpty();
     }
 
-    public snapshot() {
+    snapshot() {
         this._selectionSnapshot = cloneDeep(this._selection);
     }
 
-    public destroy() {
+    destroy() {
         this._selection.clear();
         this._selectionSnapshot.clear();
         this._hasValue$.next(false);
@@ -102,7 +104,7 @@ export class SelectionManager<T extends IGridDataEntry> {
     private _updateState = (predicate: (value: T) => void, values: T[]) => {
         values.forEach(predicate);
         this._emitChangeEvent();
-    }
+    };
 
     private _emitChangeEvent() {
         this._hasValue$.next(this.hasValue());
@@ -126,7 +128,7 @@ export class SelectionManager<T extends IGridDataEntry> {
                 this._selectedToEmit.push(value);
             }
         }
-    }
+    };
 
     private _unmarkSelected = (value: T) => {
         if (this.isSelected(value)) {
@@ -136,5 +138,5 @@ export class SelectionManager<T extends IGridDataEntry> {
                 this._deselectedToEmit.push(value);
             }
         }
-    }
+    };
 }