+++ packages/apollo-angular-components/projects/angular/components/ui-password-indicator/src/ui-password-indicator.component.spec.ts
@@ -52,22 +52,16 @@ const enum Rule {
     `,
 })
 export class PasswordIndicatorTestComponent {
-    public rules = {
+    rules = {
         [Rule.length]: (value: string) => value.length > 8,
         [Rule.upperCase]: /(?=.*[A-Z])/,
         [Rule.lowerCase]: /(?=.*[a-z])/,
         [Rule.dollarOrAt]: /(?=.*[$@])/,
     };
 
-    public form: FormGroup;
+    form: FormGroup;
 
-    public hideValidRuleDefinition = false;
-
-    public setValue(value: string) {
-        const password = this.form.get('password')!;
-        password.markAsDirty();
-        password!.setValue(value);
-    }
+    hideValidRuleDefinition = false;
 
     constructor(fb: FormBuilder) {
         this.form = fb.group({
@@ -76,6 +70,12 @@ export class PasswordIndicatorTestComponent {
                 complexityValidator(this.rules, false),
             ],
         });
+    }
+
+    setValue(value: string) {
+        const password = this.form.get('password')!;
+        password.markAsDirty();
+        password!.setValue(value);
     }
 }