+++ packages/apollo-angular-components/projects/angular/directives/ui-autofocus/src/ui-autofocus.directive.ts
@@ -55,19 +55,19 @@ export class UiAutofocusDirective implements OnInit {
      *
      */
     @Input()
-    public selectionLocation: 'start' | 'end' = 'start';
+    selectionLocation: 'start' | 'end' = 'start';
 
     /**
      * The decorated `HTMLElement` reference.
      *
      */
-    public element?: HTMLElement;
+    element?: HTMLElement;
 
     private _autofocus = true;
 
     /**
-    * @ignore
-    */
+     * @ignore
+     */
     constructor(
         private _el: ElementRef,
         private _zone: NgZone,
@@ -86,7 +86,7 @@ export class UiAutofocusDirective implements OnInit {
      * Enqueues a focus event.
      *
      */
-    public enqueueFocus() {
+    enqueueFocus() {
         if (this._autofocus) {
             this._zone.runOutsideAngular(() => {
                 merge(
@@ -113,7 +113,7 @@ export class UiAutofocusDirective implements OnInit {
      * Focus the `element`.
      *
      */
-    public focus(element?: HTMLElement) {
+    focus(element?: HTMLElement) {
         if (!element) { return; }
         element.focus();
 
@@ -137,7 +137,7 @@ export class UiAutofocusDirective implements OnInit {
         }
 
         const children = el.children || el.childNodes;
-
+        // eslint-disable-next-line @typescript-eslint/prefer-for-of
         for (let i = 0; i < children.length; i++) {
             const focusable = children[i].nodeType === ELEMENT_NODE ?
                 this._getFocusableNode(children[i] as HTMLElement) :