+++ packages/apollo-angular-components/projects/angular/directives/ui-drag-and-drop-file/src/ui-drag-and-drop-file.directive.spec.ts
@@ -31,11 +31,11 @@ import { UiDragAndDropFileDirective } from './ui-drag-and-drop-file.directive';
     `,
 })
 class TestDragAndDropFileComponent {
-    public disabled = false;
-    public multiple = false;
+    disabled = false;
+    multiple = false;
 
-    public files?: File[];
-    public fileType?: string;
+    files?: File[];
+    fileType?: string;
 
     onFileChange(files: FileList) {
         this.files = Array.from(files);
@@ -111,9 +111,18 @@ describe('Directive: UiDragAndDropFileDirective', () => {
         });
 
         [
-            { name: 'click', obj: EventGenerator.click },
-            { name: 'Keydown.Enter', obj: EventGenerator.keyDown(Key.Enter) },
-            { name: 'Keydown.Space', obj: EventGenerator.keyDown(Key.Space) },
+            {
+                name: 'click',
+                obj: EventGenerator.click,
+            },
+            {
+                name: 'Keydown.Enter',
+                obj: EventGenerator.keyDown(Key.Enter),
+            },
+            {
+                name: 'Keydown.Space',
+                obj: EventGenerator.keyDown(Key.Space),
+            },
         ].forEach(event => {
             it(`should trigger browse when ${event.name} is triggered on the parent container`, () => {
                 const spy = spyOn(fileInput, 'click');
@@ -230,7 +239,10 @@ describe('Directive: UiDragAndDropFileDirective', () => {
                 fixture.detectChanges();
 
                 dropEvent.dataTransfer.files.add(
-                    ...component.fakeFiles({ accepted: 3, rejected: 5 }),
+                    ...component.fakeFiles({
+                        accepted: 3,
+                        rejected: 5,
+                    }),
                 );
 
                 container.dispatchEvent(dropEvent);
@@ -253,12 +265,12 @@ describe('Directive: UiDragAndDropFileDirective', () => {
         describe('via input', () => {
             let changeEvent: Event;
             let targetElement: {
-                files: FakeFileList,
+                files: FakeFileList;
             };
             beforeEach(() => {
                 changeEvent = EventGenerator.change();
                 targetElement = changeEvent.target as any as {
-                    files: FakeFileList,
+                    files: FakeFileList;
                 };
             });
 
@@ -301,7 +313,6 @@ describe('Directive: UiDragAndDropFileDirective', () => {
                 expect(component.files!.length).toBe(1);
             });
 
-
             it('should NOT emit files of different types', () => {
                 targetElement.files.add(
                     ...component.fakeFiles({ rejected: 1 }),
@@ -317,7 +328,10 @@ describe('Directive: UiDragAndDropFileDirective', () => {
                 fixture.detectChanges();
 
                 targetElement.files.add(
-                    ...component.fakeFiles({ accepted: 3, rejected: 5 }),
+                    ...component.fakeFiles({
+                        accepted: 3,
+                        rejected: 5,
+                    }),
                 );
 
                 fileInput.dispatchEvent(changeEvent);