+++ packages/apollo-angular-components/projects/angular/utilities/src/rxjs/concat-join.ts
@@ -10,8 +10,8 @@ import {
 
 import { ObservedValueOf } from './internal/observed-value-of';
 
-// tslint:disable: max-line-length
-// tslint:disable: valid-js-doc
+/* eslint-disable max-len */
+/* eslint-disable  */
 /**
  * @ignore
  */
@@ -36,8 +36,8 @@ export function concatJoin<O1 extends ObservableInput<any>, O2 extends Observabl
  * @ignore
  */
 export function concatJoin<O1 extends ObservableInput<any>, O2 extends ObservableInput<any>, O3 extends ObservableInput<any>, O4 extends ObservableInput<any>, O5 extends ObservableInput<any>, O6 extends ObservableInput<any>>(v1: O1, v2: O2, v3: O3, v4: O4, v5: O5, v6: O6): Observable<[ObservedValueOf<O1>, ObservedValueOf<O2>, ObservedValueOf<O3>, ObservedValueOf<O4>, ObservedValueOf<O5>, ObservedValueOf<O6>]>;
-// tslint:enable: max-line-length
-// tslint:enable: valid-js-doc
+/* eslint-enable max-len */
+/* eslint-enable */
 
 /**
  * Captures the first emitted value from the provided stream sequence and returns an array with the first emitted value from each.