+++ packages/apollo-angular-components/projects/playground/src/app/pages/grid/grid.intl.ts
@@ -3,7 +3,7 @@ import { UiGridIntl } from '@uipath/angular/components/ui-grid';
 
 @Injectable()
 export class UiGridTable extends UiGridIntl {
-    public noDataMessageAlternative = (searchValue?: string, activeFilters?: number | null) => {
+    noDataMessageAlternative = (searchValue?: string, activeFilters?: number | null) => {
         if (searchValue && activeFilters) {
             return `No results with current filters applied for: ${searchValue}`;
         }
@@ -15,5 +15,5 @@ export class UiGridTable extends UiGridIntl {
         }
 
         return 'No results to display';
-    }
+    };
 }